fix notification event

+++ app/Traits/CanSendNotifications.php
@@ -1,8 +1,6 @@
 <?php
 namespace App\Traits;
 
-use App\Models\Product;
-use App\Models\User;
 use App\Notifications\CreateNotification;
 use App\Notifications\DeleteNotification;
 use App\Notifications\UpdateNotification;
@@ -17,16 +15,22 @@ trait CanSendNotifications
     protected static function bootCanSendNotifications(): void
     {
         static::created(function ($model) {
-            [ $users, $initiator, $feature ] = getNotificationData($model);
-            Notification::send($users, new CreateNotification($initiator, $feature));
+            if (config('app.env') == 'production') {
+                [ $users, $initiator, $feature ] = getNotificationData($model);
+                Notification::send($users, new CreateNotification($initiator, $feature));
+            }
         });
         static::updated(function ($model) {
-            [ $users, $initiator, $feature ] = getNotificationData($model);
-            Notification::send($users, new UpdateNotification($initiator, $feature));
+            if (config('app.env') == 'production') {
+                [ $users, $initiator, $feature ] = getNotificationData($model);
+                Notification::send($users, new UpdateNotification($initiator, $feature));
+            }
         });
         static::deleting(function ($model) {
-            [ $users, $initiator, $feature ] = getNotificationData($model);
-            Notification::send($users, new DeleteNotification($initiator, $feature));
+            if (config('app.env') == 'production') {
+                [ $users, $initiator, $feature ] = getNotificationData($model);
+                Notification::send($users, new DeleteNotification($initiator, $feature));
+            }
         });
     }
 }
